truth-engine/requirements.txt
fastapi
uvicorn[standard]
groq
langgraph
langchain-core
beautifulsoup4
httpx
python-dotenv
sse-starlette
pydantic
python-multipart
